refactor/fix: refatora services e DTOs para implementação do BigDecimal

+++ src/main/java/br/com/centroweg/financas/domain/interfaces/CalculoImpostoStrategy.java
@@ -2,7 +2,9 @@ package br.com.centroweg.financas.domain.interfaces;
 
 import br.com.centroweg.financas.domain.entities.ativo.Ativo;
 
+import java.math.BigDecimal;
+
 public interface CalculoImpostoStrategy {
-    Double calcular(Double valor);
+    BigDecimal calcular(BigDecimal valor);
     boolean isAplicavel(Ativo ativo);
 }
