+++ src/main/java/br/com/centroweg/financas/service/dto/ativo/AtivoResponseDTO.java
@@ -10,6 +10,6 @@ public record AtivoResponseDTO(
         String tipo,
         Double risco,
         String detalhesAtivo,
-        Double impostoEstimado
+        BigDecimal impostoEstimado
 ) {
 }
