+++ src/main/java/br/com/centroweg/financas/service/dto/investimento/HistoricoPrecosDTO.java
@@ -1,8 +1,10 @@
 package br.com.centroweg.financas.service.dto.investimento;
 
+import java.math.BigDecimal;
+
 public record HistoricoPrecosDTO (
         String ticker,
-        Double preco,
+        BigDecimal preco,
         String dataRegistro
 ){
 }
